% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{BesselK}
\alias{BesselK}
\title{Modified Bessel function of the second kind}
\usage{
BesselK(nu, z)
}
\arguments{
\item{nu}{a real positive value}

\item{z}{a real positive value}
}
\description{
This function calls the GSL scientific library to evaluate 
the modified Bessel function of the second kind.
}
\seealso{
\code{\link{matern}}
}
\author{
Pulong Ma \email{mpulong@gmail.com}
}
