% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/AllClass.R
\docType{methods}
\name{show,gp-method}
\alias{show,gp-method}
\alias{show,gp-methods}
\title{Print the information an object of the \code{\link{gp}} class}
\usage{
\S4method{show}{gp}(object)
}
\arguments{
\item{object}{an object of \code{\link{gp}} class}
}
\description{
Print the information an object of the \code{\link{gp}} class
}
