% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaSPUtils.R
\name{gp.condsim}
\alias{gp.condsim}
\title{Perform conditional simulation from a Gaussian process}
\usage{
gp.condsim(obj, XX, nsample = 1, seed = NULL)
}
\arguments{
\item{obj}{an \code{S4} object \linkS4class{gp}}

\item{XX}{a matrix of new locations where conditional simulation is performed.}

\item{nsample}{number of conditional simulations default at 1}

\item{seed}{random generation seed default to be \code{NULL}.}
}
\value{
an array (vector or matrix) of conditional simulations
}
\description{
This function simulate from the Gaussian process model conditional on existing 
data (i.e., locations, observations). This is known as conditional simulation in geostatistics, which 
simulates realizations from the (posterior) predictive distribution of the process given the data.
}
