% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaSPUtils.R
\name{gp.get.mcmc}
\alias{gp.get.mcmc}
\title{get posterior summary for MCMC samples}
\usage{
gp.get.mcmc(obj, burnin = 500)
}
\arguments{
\item{obj}{a \code{\link{gp}} object}

\item{burnin}{a numerical value specifying the burn-in period for calculating posterior summaries.}
}
\description{
This function processes posterior samples in the \code{\link{gp}} object.
}
\seealso{
\link{GPBayes-package}, \code{\link{GaSP}}, \code{\link{gp}}, \code{\link{gp.mcmc}}
}
