% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GaSPUtils.R
\name{gp.optim}
\alias{gp.optim}
\title{A wraper to fit a Gaussian stochastic process model with optimization methods}
\usage{
gp.optim(obj, method = "MMLE", opt = NULL)
}
\arguments{
\item{obj}{an \code{S4} object \linkS4class{gp}}

\item{method}{a string indicating the parameter estimation method:
\describe{
\item{MPLE}{This indicates that the \emph{maximum profile likelihood estimation} 
(\strong{MPLE}) is used.}
\item{MMLE}{This indicates that the \emph{maximum marginal likelihood estimation} 
(\strong{MMLE}) is used.}
\item{MAP}{This indicates that the marginal/integrated posterior is maximized. It's not yet supported.}
}}

\item{opt}{a list of arguments to setup the \code{\link[stats]{optim}} routine. Current implementation uses three arguments: 
\describe{
 \item{method}{The optimization method: \code{Nelder-Mead} or \code{L-BFGS-B}.}
\item{lower}{The lower bound for parameters.}
\item{upper}{The upper bound for parameters.}
}}
}
\value{
a list of updated \code{\link{gp}} object \code{obj} and 
fitted information \code{fit}
}
\description{
This function is a wraper to estimate parameters in the GaSP model with different
choices of estimation methods using numerical optimization methods.
}
\examples{
 
code = function(x){
y = (sin(pi*x/5) + 0.2*cos(4*pi*x/5))*(x<=9.6) + (x/10-1)*(x>9.6) 
return(y)
}
n=100
input = seq(0, 20, length=n)
XX = seq(0, 20, length=99)
Ztrue = code(input)
set.seed(1234)
output = Ztrue + rnorm(length(Ztrue), sd=0.1)
obj = gp(formula=~1, output, input, 
        param=list(range=4, nugget=0.1,nu=2.5),
        smooth.est=FALSE,
        cov.model=list(family="matern", form="isotropic"))
        
fit.optim = gp.optim(obj, method="MPLE")
fit.optim = gp.optim(obj, method="MPLE", 
                    opt=list(method="L-BFGS-B"))

obj = gp(formula=~1, output, input, 
        param=list(range=4,tail=0.5, nugget=0.1,nu=.5),
        smooth.est=TRUE,
        cov.model=list(family="CH", form="isotropic"))

fit2 = gp.optim(obj, method="MPLE", 
                      opt=list(
                      method="L-BFGS-B",
                      upper=c(1e10, 5, 5, 3)))
}
\seealso{
\link{GPBayes-package}, \code{\link{GaSP}}, \linkS4class{gp}, \code{\link{gp.mcmc}}
}
\author{
Pulong Ma \email{mpulong@gmail.com}
}
