% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{HypergU}
\alias{HypergU}
\title{Confluent hypergeometric function of the second kind}
\usage{
HypergU(a, b, x)
}
\arguments{
\item{a}{a real value}

\item{b}{a real value}

\item{x}{a real value}
}
\description{
This function calls the GSL scientific library to evaluate 
the confluent hypergeometric function of the second kind; see Abramowitz and Stegun 1972, p.505.
}
\seealso{
\code{\link{CH}}
}
\author{
Pulong Ma \email{mpulong@gmail.com}
}
