% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/util.R
\name{update_list}
\alias{update_list}
\title{Add elements of sup_list to sub_list}
\usage{
update_list(sup_list, sub_list)
}
\arguments{
\item{sup_list}{high-priority list (will overwrite sub_list values)}

\item{sub_list}{low-priority list (will be overwritten by sup_list values)}
}
\value{
combine sub_list and sup_list. Any element that is in both sup_list
and sub_list will be overwritten by sup_list
}
\description{
Add elements of sup_list to sub_list
}
\examples{
const <- list(x = 1)
my_const <- list(y = 3)
update_list(my_const, const)
}
